Nick and Adam got the local storage functionality to be able to display all of the saved cards from local storage back onto the deck.html page

+++ index.js
@@ -84,6 +84,9 @@ var pokeCard = function (cards) {
 
         });
 };
+
+let pokemonInfo;
+
 var displayCard = function (cardImg) {
     //for (var i = 0; i < cardImg.length; i++) {
     console.log("card img", cardImg[0].images.large)
@@ -92,25 +95,27 @@ var displayCard = function (cardImg) {
     pokeCardImg.src = (cardImg[0].images.large);
     tradingCard.appendChild(pokeCardImg);
 
-    const pokemonInfo = {
+    pokemonInfo = {
         id: cardImg[0].id,
         name: cardImg[0].name,
         imageUrl: cardImg[0].images.large
     }
 
     console.log(cardImg[0])
-    saveButton(pokemonInfo);
+
     /// create a href for link to image
     ///
 }
 
 var saveButton = function (pokemonInfo) {
-    document.getElementById("save-card").addEventListener("click", function() {
-        const existingPokemon = JSON.parse(localStorage.getItem("pokemonDeck")) || [];
-        existingPokemon.push(pokemonInfo);
-        localStorage.setItem("pokemonDeck", JSON.stringify(existingPokemon));
-    });
-    
+    const existingPokemon = JSON.parse(localStorage.getItem("pokemonDeck")) || [];
+
+    // make sure that we don't add any duplicate cards to the page where we view all saved cards
+    const existingCard = existingPokemon.find(currentPokemon => currentPokemon.id === pokemonInfo.id);
+    if (existingCard) { return;}
+    existingPokemon.push(pokemonInfo);
+    localStorage.setItem("pokemonDeck", JSON.stringify(existingPokemon));
+
 }
 
 // view deck button
@@ -121,3 +126,5 @@ viewDeck.addEventListener("click", function () {
  
 // add event listeners to forms
 userFormEl.addEventListener('submit', formSubmitHandler);
+
+document.getElementById("save-card").addEventListener("click", function(){saveButton(pokemonInfo)});
